Gui: move CommandCompleter into its own source file
