+++ src/Gui/CMakeLists.txt
@@ -396,6 +396,7 @@ SET(Command_CPP_SRCS
     CommandLink.cpp
     CommandPyImp.cpp
     ShortcutManager.cpp
+    CommandCompleter.cpp
 )
 SET(Command_SRCS
     ${Command_CPP_SRCS}
@@ -404,6 +405,7 @@ SET(Command_SRCS
     Command.h
     CommandT.h
     ShortcutManager.h
+    CommandCompleter.h
 )
 SOURCE_GROUP("Command" FILES ${Command_SRCS})
