+++ src/Gui/DlgToolbarsImp.cpp
@@ -37,6 +37,7 @@
 #include "Application.h"
 #include "BitmapFactory.h"
 #include "Command.h"
+#include "Action.h"
 #include "ToolBarManager.h"
 #include "MainWindow.h"
 #include "Widgets.h"
@@ -206,7 +207,8 @@ void DlgCustomToolbars::importCustomToolbars(const QByteArray& name)
                 if (pCmd) {
                     // command name
                     QTreeWidgetItem* item = new QTreeWidgetItem(toplevel);
-                    item->setText(0, qApp->translate(pCmd->className(), pCmd->getMenuText()));
+                    item->setText(0, Action::commandMenuText(pCmd));
+                    item->setToolTip(0, Action::commandToolTip(pCmd));
                     item->setData(0, Qt::UserRole, QByteArray(it2->first.c_str()));
                     if (pCmd->getPixmap())
                         item->setIcon(0, BitmapFactory().iconFromTheme(pCmd->getPixmap()));
